+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - NOTE: There is an inconsistency in the order of application of `ruby2_keywords` and `instrument_method` in
   Ruby >= 3.2. We recommend not combining these two annotations at all, but if you must, call `instrument_method`
   after your method has been defined.
+- [IMPROVEMENT] When Rails's `exceptions_app` (used by ActionDispatch::ShowExceptions) is set to another instrumented
+  responder (like a Rails router), Skylight will not set the endpoint name after exception processing has started.
+  This means that error traces will now be aggregated under the original endpoint that generated the error (but with
+  the 'error' segment), rather than under the exception handler's controller and action name.
 
 ## 5.3.4 (October 17, 2022)